// src/app/modules/shared/model/export-resultat-bureau-vote.ts
export class ExportResultatBureauVote {
	codeLieuVote: string;
	lieuVote: string;
	BureauVote: string;
	inscrits: number;
	nombreHommeVotant: number;
	nombreFemmeVotant: number;
	totalVotant: number;
	tauxParticipation: number;
	nombreBulletinNull: number;
	nombreBulletinBlanc: number;
	suffrageExprimes: number;
	candidat: string;
	score: number;
}